Fix compiler warning C4242: 'argument' : conversion from 'int' to 'WORD', possible loss of data.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@7075 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/ShareazaFile.cpp
@@ -130,7 +130,7 @@ bool CShareazaFile::SplitStringToURLs(LPCTSTR pszURLs, CMapStringToFILETIME& oUr
 				! Network.IsFirewalledAddress( &nAddress, TRUE ) &&
 				! Network.IsReserved( (IN_ADDR*)&nAddress ) )
 			{
-				strURL = GetURL( *(IN_ADDR*)&nAddress, nPort );
+				strURL = GetURL( *(IN_ADDR*)&nAddress, static_cast< DWORD >( nPort ) );
 			}
 			else
 			{
